pyproject.toml
[project]
name = "lightworld"
version = "0.1.0"
description = "LightWorld - 多模态社会模拟与推演引擎"
requires-python = ">=3.11"
license = { text = "AGPL-3.0" }
authors = [
    { name = "LightWorld Team" }
]

dependencies = [
    # 核心框架
    "flask>=3.0.0",
    "flask-cors>=6.0.0",

    # LLM 相关
    "openai>=1.0.0",

    # Zep Cloud
    "zep-cloud==3.13.0",

    # OASIS 社交媒体模拟
    "camel-oasis==0.2.5",
    "camel-ai==0.2.78",

    # 文件处理
    "PyMuPDF>=1.24.0",
    "Pillow>=10.0.0",
    # 编码检测（支持非UTF-8编码的文本文件）
    "charset-normalizer>=3.0.0",
    "chardet>=5.0.0",

    # 工具库
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "socksio>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pipreqs>=0.5.0",
]

[project.scripts]
lightworld-api = "lightworld.cli.api:main"
lightworld-local-pipeline = "lightworld.cli.local_pipeline:main"
lightworld-parallel-sim = "lightworld.cli.parallel_simulation:main"
lightworld-full-run = "lightworld.cli.full_run:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/lightworld"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
